[Chore] move `animation.ts` to `src/base/browser/basic`
