+++ src/base/browser/basic/animation.ts
